Types
* major bug fix
  - basic types 'u32' and 'i32' where mapped to an incorrect [u]int*_t type
  - added a unit test to check if the size is correct!

+++ c/type.h
@@ -42,13 +42,13 @@ extern "C" {
 typedef bool u1;
 typedef uint8_t u8;
 typedef uint16_t u16;
-typedef uint16_t u32;
+typedef uint32_t u32;
 typedef uint64_t u64;
 
 typedef bool i1;
 typedef int8_t i8;
 typedef int16_t i16;
-typedef int16_t i32;
+typedef int32_t i32;
 typedef int64_t i64;
 
 #ifdef __cplusplus
